@@ -9,20 +9,20 @@
   "license": "GPL-3.0",
   "description": "UI app for Photoview",
   "dependencies": {
-    "@apollo/client": "^3.3.21",
-    "@babel/preset-typescript": "^7.14.5",
-    "@craco/craco": "^6.2.0",
-    "@headlessui/react": "^1.4.1",
-    "@react-aria/focus": "^3.4.0",
+    "@apollo/client": "^3.5.8",
+    "@babel/preset-typescript": "^7.16.7",
+    "@craco/craco": "^6.4.3",
+    "@headlessui/react": "^1.4.3",
+    "@react-aria/focus": "^3.5.0",
     "@rollup/plugin-babel": "^5.3.0",
     "@types/geojson": "^7946.0.8",
-    "@types/jest": "^26.0.24",
-    "@types/mapbox-gl": "^2.3.3",
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-helmet": "^6.1.2",
-    "@types/react-router-dom": "^5.1.8",
-    "@types/styled-components": "^5.1.11",
+    "@types/jest": "^27.4.0",
+    "@types/mapbox-gl": "^2.6.0",
+    "@types/react": "^17.0.38",
+    "@types/react-dom": "^17.0.11",
+    "@types/react-helmet": "^6.1.5",
+    "@types/react-router-dom": "^5.3.3",
+    "@types/styled-components": "^5.1.21",
     "@types/url-join": "^4.0.1",
     "autoprefixer": "^9.8.6",
     "babel-plugin-graphql-tag": "^3.3.0",
@@ -30,28 +30,28 @@
     "connect-history-api-fallback": "^1.6.0",
     "copy-to-clipboard": "^3.3.1",
     "eslint-config-prettier": "^8.3.0",
-    "eslint-plugin-jest-dom": "^3.9.0",
+    "eslint-plugin-jest-dom": "^4.0.1",
     "fs-extra": "^10.0.0",
-    "i18next": "^20.3.2",
-    "mapbox-gl": "^2.3.1",
+    "i18next": "^21.6.10",
+    "mapbox-gl": "^2.6.1",
     "postcss": "^7.0.36",
-    "prettier": "^2.3.2",
-    "prop-types": "^15.7.2",
+    "prettier": "^2.5.1",
+    "prop-types": "^15.8.1",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-helmet": "^6.1.0",
-    "react-hook-form": "^7.11.0",
-    "react-i18next": "^11.11.1",
-    "react-router-dom": "^5.2.0",
+    "react-hook-form": "^7.25.3",
+    "react-i18next": "^11.15.3",
+    "react-router-dom": "^6.2.1",
     "react-router-prop-types": "^1.0.5",
     "react-scripts": "^4.0.3",
     "react-spring": "^8.0.27",
     "react-test-renderer": "^17.0.2",
-    "styled-components": "^5.3.0",
-    "subscriptions-transport-ws": "^0.9.19",
-    "tailwind-override": "^0.2.3",
-    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.4",
-    "typescript": "^4.3.5",
+    "styled-components": "^5.3.3",
+    "subscriptions-transport-ws": "^0.11.0",
+    "tailwind-override": "^0.6.1",
+    "tailwindcss": "npm:@tailwindcss/postcss7-compat@^2.2.17",
+    "typescript": "^4.5.5",
     "url-join": "^4.0.1"
   },
   "scripts": {
@@ -69,15 +69,15 @@
     "prepare": "(cd .. && npx husky install)"
   },
   "devDependencies": {
-    "@testing-library/jest-dom": "^5.14.1",
-    "@testing-library/react": "^12.0.0",
-    "@testing-library/user-event": "^13.1.9",
-    "apollo": "2.33.4",
-    "apollo-language-server": "1.26.3",
-    "husky": "^6.0.0",
-    "i18next-parser": "^4.2.0",
-    "lint-staged": "^11.0.1",
-    "tsc-files": "^1.1.2"
+    "@testing-library/jest-dom": "^5.16.1",
+    "@testing-library/react": "^12.1.2",
+    "@testing-library/user-event": "^13.5.0",
+    "apollo": "2.33.9",
+    "apollo-language-server": "1.26.7",
+    "husky": "^7.0.4",
+    "i18next-parser": "^5.4.0",
+    "lint-staged": "^12.3.2",
+    "tsc-files": "1.1.2"
   },
   "prettier": {
     "trailingComma": "es5",
